extensive drag'n'drop study. short answer: it's crap.
long answer: you can't change or hide the cursor during a drag event, meaning you have either a forbidding "not allowed" circle-with-a-line, a cursor with a box and a plus, a cursor with only a box, or a mix of them. in the end, drawing on the canvas has hardly any connection with drawing on a canvas and mousedown, mouseup, mouseleave, and mouseenter events are the easiest way to go - instead of undoing all the features that drag'n'drop brings with.